This is a combination of 20 commits
Added the basic structure for the new layered page

Created a new layout to hold the new layered view.  It uses parallax on
the background image, and offers a way to add a footer widget.

Added the banner content

Added banner content for the OBS page including the logo.

Added the search bar for the translations

Added a search bar for the translations, and added the styles for it.

Added search functionality

Added the search functionality to the OBS stories page.

Moved the obs includes into their own folders for better organization.

Pulling translations from both api feeds.

Completed the OBS page.

Added links on the available translations to the language pages.

Added links for all language boxes.

Creator the base generator for creating the language pages on the fly.

Moved the API call to the catalog.

Added accordion to the language page view.

Added resources to language pages.

Added correct links to open the web bible to right translation.

Added lang and direction to the appropriate areas of the page.

Moved the template into the layouts folder.

Darkened the text colors.

Change language page links to be relative.

Added USFM and Github links for ULB and UDB

Merge cleanup