Restructured code for language page generator.
Since we are adding several new items to the OBS section of the language
page,  the code was going to be difficult to manage.  I cleaned up code,
and added separate classes to handle the parsing and collecting of data for the page
generator.

Added a check to verify the Parser class is available.